styles: Improve project cards
Made the card a link and hover on click.

There is still a bug whenever the arrows are clicked when
a card's hover transition is still playing

Extracted external link to an include and stopped suing it
in the cards.

Made constant variables for css values based on imported
theme colors