feat: Final changes remove junkiness
Made sure there is no margin at the top but gave each
fullscreen section a padding-top, which fixes the offset
with the header without causing scroll overflow outside
the scroll container.

Removed the scrollbar from the main scrolling of the
home page.

Small improvement making the vertical divider thicker
to avoid it flickering.