styles: Improve card behaviour in skills
Also remove the media breakpoints which
set less lines till ellipsis
in project ones.

Project card content can be hidden because
each is linked to the project's page,
where the user can see the full content.

Skills however need to always show the full
content as that's the only place
the user can read this info on.