fix: Render single image without carousel wrapper
Before whenever we had a single image for a project page,
it was not properly positioned when wrapped
unecessarily inside 'project-image-carousel' container.

Now we only do the logic of a single image in the project
layout instead of the general image carousel component,
to ensure we are only wrapping the
carousel of images in the container
and not when its just one image.