style: Replace project link icon to better one
Instead of using the img icon with an icon
that is big and represents a link which opens
to a new tab (which is not the case),
we use the UTF encoding for the link character
which fits better and takes less place in smaller
screens.

Made sure that when the text is hidden, the
symbol is still shown.